+++ src/util/Badges.hpp
@@ -18,11 +18,9 @@ public:
         CrewManager = 2,
         CrewMember = 3,
         TeamManager = 4,
-        TeamMember = 4
+        TeamMember = 5
     };
 
     static std::map<BadgeID, std::string> getBadgeStringID;
     static std::map<std::string, std::string> getBadgeSpriteName;
-
-    void setNewBadge(std::string id, cocos2d::CCLayer* mLayer, float size, auto parent);
 };